requirements.txt
streamlit>=1.28.0
pinecone>=6.0.2
sentence-transformers>=2.2.2
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pymongo>=4.5.0
requests>=2.31.0
beautifulsoup4>=4.12.2
tqdm>=4.66.1
argparse>=1.4.0
mcp[cli]==1.6.0 
openai==1.75.0
python-dotenv
ipykernel
httpx